Add analyze and stem methods (analogs to suggest)
The "analyze" and "stem" functions in Hunspell can use almost exactly
the same code as the suggest method that is currently implemented in
HunspellJNA. I just copied the code from suggest and changed the names
of variables and the function to call in Hunspell.

These functions are useful for grammar checks and parsing.